Library/Collision: Implement SensorConnector
