+++ lib/al/Library/Collision/PartsConnectorUtil.cpp
@@ -6,7 +6,7 @@
 #include "Library/Collision/KCollisionServer.h"
 #include "Library/Collision/PartsConnector.h"
 #include "Library/Collision/PartsMtxConnector.h"
-#include "Library/Collision/SensorConnector.h"
+#include "Library/Collision/PartsSensorConnector.h"
 #include "Library/LiveActor/ActorModelFunction.h"
 #include "Library/LiveActor/ActorPoseUtil.h"
 #include "Library/LiveActor/LiveActor.h"
